Add a passing test (red/green/refactor)
This time when the test was added, it's already passing.
This is fine - it means the code already handles that case.
I'll leave the test to make sure it continues to handle the case right.
I also don't see anything that needs refactoring at this point so
this round is all done after adding the new test.